+++ test/serde/msgpack/test_msgpack_serde.py
@@ -20,24 +20,7 @@ from syft.serde.syft_serializable import SyftSerializable
 from syft.serde.msgpack import native_serde
 from syft.serde.msgpack import torch_serde
 from syft.workers.virtual import VirtualWorker
-
-
-class Experiment1(SyftSerializable):
-    def __init__(self, value):
-        self.value = value
-
-    @staticmethod
-    def simplify(worker, obj):
-        return obj.value
-
-    @staticmethod
-    def detail(worker, obj):
-        return Experiment1(obj)
-
-    @staticmethod
-    def get_msgpack_code():
-        return {"code": 12345}
-
+from test.serde.serde_helpers import SerializableDummyClass
 
 def test_tuple_simplify(workers):
     """This tests our ability to simplify tuple types.
@@ -885,16 +868,19 @@ def test_no_simplifier_found(workers):
 
 
 def test_external_lib_msgpack():
-    example1 = Experiment1("test")
+    example1 = SerializableDummyClass("test")
     ser1 = syft.serde.serialize(example1)
     result1 = syft.serde.deserialize(ser1)
     assert example1.value == result1.value
 
-    delattr(Experiment1, "get_msgpack_code")
+    saved_attr = SerializableDummyClass.get_msgpack_code
+    delattr(SerializableDummyClass, "get_msgpack_code")
     syft.serde.msgpack.serde.msgpack_global_state.stale_state = True
+
     with pytest.raises(Exception) as e:
-        example2 = Experiment1("test")
+        example2 = SerializableDummyClass("test")
         _ = syft.serde.serialize(example2)
 
     assert isinstance(e.value, syft.exceptions.UndefinedProtocolTypeError) == 1
-    setattr(example1, "get_msgpack_code", lambda: {"code": 12345})
+    setattr(SerializableDummyClass, "get_msgpack_code", saved_attr)
+    syft.serde.msgpack.serde.msgpack_global_state.stale_state = True
